Moving types to own file
