+++ frontend/components/deployment/deployment-details.tsx
@@ -9,7 +9,7 @@ import {
   TooltipProvider,
   TooltipTrigger,
 } from '@/components/ui/tooltip';
-import { Deployment } from '@/components/deployment/deployments-dashboard';
+import { Deployment } from '@/models/IDeployment';
 import { useRouter } from 'next/navigation';
 import { useEffect, useState } from 'react';
 import Link from 'next/link';
