+++ frontend/components/deployment/deployments-dashboard.tsx
@@ -5,52 +5,7 @@ import { useState } from 'react';
 import Sidebar from '../main/sidebar';
 import DeploymentCard from './deployment-card';
 import { Button } from '@/components/ui/button';
-
-export interface Deployment {
-  id: string;
-  name: string;
-  letter: string;
-  color: string;
-  status: string;
-  address?: {
-    street: string;
-    suite: string;
-    city: string;
-    zipCode: string;
-  };
-  contacts?: {
-    name: string;
-    email: string;
-  }[];
-  services: Service[];
-  compliance?: {
-    gdpr: 'compliant' | 'warning' | 'non-compliant';
-    ccpa: 'compliant' | 'warning' | 'non-compliant';
-  };
-}
-
-export interface Service {
-  name: string;
-  status: string;
-  type: string;
-  versions: Version[];
-  devices?: Device[];
-}
-
-export interface Version {
-  number: string;
-  color: 'green' | 'yellow' | 'red';
-}
-
-export interface Device {
-  name: string;
-  id: string;
-  version: string;
-  versionColor: string;
-  rootHash: string;
-  lastAccessed: string;
-  lastHeartbeat: string;
-}
+import { Deployment } from '@/models/IDeployment';
 
 export default function DeploymentsDashboard() {
   const [deployments] = useState<Deployment[]>([
